backend/requirements.txt
fastapi
uvicorn[standard]==0.22.0
groq
python-dotenv